feat : build custom hook for getting user location
